Edit renderTask parameter, Array in global scope
removed parameters from it since the array it's
using is in the global scope.

Added .type = "button"; to prevent
submission.

Added Sort.States to each filterbutton.

As it stands the "Active" filter works correctly.
The "completed" filter shows nothing, and the
"All" filter shows the "Active" tasks. This
tells me I need to alter the arrays in some way.