Add interfaces and api folder and files
Used GPT-5 to help me establish a structure
on where things should go and what they should
be named, while getting it to explain why.

Right now each request gets a an object (Scryfall
ListResponse<T>) with another object shown in an
array (ScryfallCard). I admit this is a bit too
complex for me to create on my own, but I hope to
learn something of it.