Add card interface, DOMfunction and api direcory
Again, a lot of things here have to deep
complexity for me to understand in it's whole.

Appantly I had to await the response which was of
the type ScryfallListResponse<scryfallCard>
and then take the resul.data (Which I thought
I already did).

CoPilot already knows what the json will
(and could) give, thus helping me create working
interfaces.

Next up will be setting up the result in the
webpage.